+++ app/Providers/AppServiceProvider.php
@@ -8,8 +8,6 @@ use Illuminate\Support\ServiceProvider;
 use App\Http\Controllers\SSO\CustomAccessToken;
 use App\Models\TeamHasDataAccessApplication;
 use App\Observers\TeamHasDataAccessApplicationObserver;
-use Illuminate\Support\Facades\Http;
-use Illuminate\Support\Facades\Cache;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -62,40 +60,6 @@ class AppServiceProvider extends ServiceProvider
 
         TeamHasDataAccessApplication::observe(TeamHasDataAccessApplicationObserver::class);
 
-
-        $url = env('FEATURE_FLAGGING_CONFIG_URL');
-        if ($url) {
-            $featureFlags = Cache::remember('feature_flags', now()->addMinutes(60), function () use ($url) {
-                $res = Http::get($url);
-                if ($res->successful()) {
-                    return $res->json();
-                }
-
-                logger()->error('Failed to fetch feature flags from URL', ['url' => $url]);
-                return [];
-            });
-
-            if (is_array($featureFlags)) {
-                $this->defineFeatureFlags($featureFlags);
-            }
-        }
     }
-    protected function defineFeatureFlags(array $flags, string $prefix = '')
-    {
-        foreach ($flags as $key => $value) {
-            $fullKey = $prefix ? "{$prefix}.{$key}" : $key;
-            if (is_array($value)) {
-                if (isset($value['enabled']) && is_bool($value['enabled'])) {
-                    \Laravel\Pennant\Feature::define($fullKey, $value['enabled']);
-                    logger()->info("Feature flag defined: {$fullKey} = " . ($value['enabled'] ? 'ENABLED' : 'DISABLED'));
-                }
 
-                foreach ($value as $subKey => $subVal) {
-                    if (is_array($subVal)) {
-                        $this->defineFeatureFlags([$subKey => $subVal], $fullKey);
-                    }
-                }
-            }
-        }
-    }
 }
