+++ phpunit.xml
@@ -13,7 +13,8 @@
   </testsuites>
   <coverage/>
   <php>
-    <env name="APP_ENV" value="testing"/>
+   <env name="APP_ENV" value="testing"/>
+    <env name="FEATURE_FLAGGING_CONFIG_URL" value="https://github.com/HDRUK/hdruk-feature-configurations/blob/main/dev/features.json" />
     <env name="BCRYPT_ROUNDS" value="4"/>
     <env name="CACHE_DRIVER" value="array"/>
     <env name="DB_CONNECTION" value="sqlite"/>
